requirements-cpu.txt
flasgger==0.9.5
Flask==1.1.2
Pillow==8.1.0
typing==3.7.4.3
python-dotenv==0.15.0
opencv-python-headless==4.5.1.48
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.8.0+cpu
-f https://download.pytorch.org/whl/torch_stable.html
torchvision==0.9.0+cpu -f https://download.pytorch.org/whl/torch_stable.html
pytorch-lightning==1.2.4
pandas==1.2.3



